PR feedback (low hanging fruit first)
