+++ crates/rome_js_syntax/src/lib.rs
@@ -8,6 +8,7 @@ pub mod binding_ext;
 pub mod directive_ext;
 pub mod expr_ext;
 pub mod file_source;
+pub mod function_ext;
 pub mod identifier_ext;
 pub mod import_ext;
 pub mod jsx_ext;
@@ -24,6 +25,7 @@ mod union_ext;
 pub use self::generated::*;
 pub use expr_ext::*;
 pub use file_source::*;
+pub use function_ext::*;
 pub use identifier_ext::*;
 pub use modifier_ext::*;
 pub use rome_rowan::{
