+++ npm/rome/configuration_schema.json
@@ -811,7 +811,7 @@
 					]
 				},
 				"noExcessiveComplexity": {
-					"description": "The more complexity a function contains, the harder it is to understand later on.",
+					"description": "Disallow functions that exceed a given complexity score.",
 					"anyOf": [
 						{ "$ref": "#/definitions/RuleConfiguration" },
 						{ "type": "null" }
