+++ website/src/pages/lint/rules/index.mdx
@@ -948,8 +948,7 @@ Prevents JSX properties to be assigned multiple times.
 <h3 data-toc-exclude id="noExcessiveComplexity">
 	<a href="/lint/rules/noExcessiveComplexity">noExcessiveComplexity</a>
 </h3>
-The more complexity a function contains, the harder it is to understand
-later on.
+Disallow functions that exceed a given complexity score.
 </section>
 <section class="rule">
 <h3 data-toc-exclude id="noForEach">
